Fix divider missing with AppButtonsPreference
Symptom:
The divider below of the AppButtonsPreference is missing on the
Storage & cache and Battery pages of app.

Root cause:
The PreferenceCategory divider has been hidden for Android S design.
If we put the AppButtonsPreference and the PreferenceCategory together,
the divider will be default hidden by PreferenceCategory.

Solution:
To add back the divider for AppButtonsPreference, override the
allowDividerAbove in the PreferenceCategory.

Fixes: 179118594
Test: visual
Change-Id: I74ba7679a69a578a87734b9920f53a20c79d6567