Update FingerprintEnrollEnrolling for UDFPS
The BC theme didn't work in the fingerprint enrollment page since this
page was using a customized layout and wasn't following the SUD
template. Also the fingerprint sensor icon has been moved to SysUI so
it's unnecessary to have the customized layout.

This CL is trying to merge two layouts together and make BC theme apply
to the fingerprint enrollment page.

Bug: 177026664
Test: visual verified
Change-Id: Ia22ea14244cd4b508a1fa6341aa15bd741c195f4