Fix overlapping problem for device admin page
The content of Activate device admin app is overlapping with the title
while scrolling up. That is because the collapsing toolbar needs to
work with the view that supports nested scrolling feature, and this page
hasn't updated the layout to use NestedScrollView. So that caused the
overlapping problem.

Fix: 188184864
Test: visual verified
1) Navigate to Settings > Security > Device admin apps > pick up one to
activate
2) Scrolling the content and see if the content is overlapping with the
title

Change-Id: Ide5ff9ed40202c3b217934db93c82a4861983416