Fix the link text "Learn more" in the accessibility pages with help links could not convey the link's purpose clearly
Root cause: A long content textview shows the accessibility service info and learn more link. User cannot double click to open help link. They need to swipe up and right to pops up dialog to move to next page. But, the dialog show the "Learn more", it is not clear for user to idendify what the link page.

Solution: Add another textview to show the "Learn more" and provides more clear content description for this link. Also, user just need to double click to activate it.

Bug: 188603037
Test: make RunSettingsRoboTests -j52 ROBOTEST_FILTER=AccessibilityFooterPreferenceTest
Change-Id: Ia9a56f11bbd235ae3a698fcb2e1a2fcff64ef940