Turn off transcoding notification by default
To aid debugging, we included transcoding progress
notifications. We don't intend to launch with the notifications on by
default so turn it off now in preparation for Android S launch.

The notifications can still be enabled in the developer options with:
Settings > System > Developer options > Media transcoding >
Show transcoding notifications

Test: Manual
Bug: 190562454
Change-Id: Icab62cb91fdf22e5308d0c4bbc710b88a51170d3