Revert "[Settings] Code Refactor - ActiveSubscip..."
There seems to be an issue on system when a modification on file name get referenced by multiple submit(s).
(ag/15106410 and ag/15102785).

It seems to me that, build system got build breakage immediately instead of merging all of the files (even within same package).

Reason for revert: Try to submit all of the modifications within single submit as a workaround.

Bug: 192204057
Test: build pass
Change-Id: Ie497088a58c6ac564fde193fad2c0b67a965933c