Cosmetic changes to zen mode people & sound/vibration screens
On the people screen, have the footer show up all the time and provide a more relevant message (indicating that the chosen people override any app/etc settings) than the current setting.

On the sound/vibration screen, remove the sub-title since it's a bit confusing what exactly is being turned on or off.

Bug: 190180868
Bug: 215564123
Test: manual
Change-Id: I78d407af350a9f99d26d1e5082b752881bd3f186