Fix quick settings tooltip changed when color correction fab clicked
Root cause: Users click on color correction fab to make foreground activity recreate. When recreated, the ag/13472947 makes the MainSwitch updates to ON status and get the callback for #onPreferenceToggled. It pops up the direct-use style tooltip rather than the edit style tooltip.
Soluton: Reference #updateSwitchBarToggleSwitch to early returns if values are same.

Bug: 229167631
Test: Manual testing
Change-Id: I6781fd509db7538898ad6ff0ae47124c9b9e1bc5