Cancel bonding upon back gesture.
When pairing dialog is dismissed using the back gesture, we do not call
cancelBondProcess() for the remote device. Since bonding is not
cancelled and the dialog is dismissed, when the user tries to pair again
and scans for devices, nothing shows up. This is because in case of a
pending bond or pending SDP, the request to search for new devices is
queued.

The correct behavior is to cancel bonding if the pairing dialog is
dismissed using the back gesture. This is similar to what
happens when the pairing dialog is dismissed using the cancel button.

Bug: 231554812
Test: Changed settings app is able to scan after dismissing the dialog

Change-Id: Ia790e345be811be1b60762ff819544d03c5a18fd
Merged-In: Ia790e345be811be1b60762ff819544d03c5a18fd