Apply the glifv3 theme when the theme string in intentExtra is glifv4.
Create the short-term solution for the issue in settings page.

The issue is because when we set to use glifv4 theme, the theme util in settings do not have that version(glifv4) info in their util so that they will roll back to the glif v1 theme.

Current solution is rollback to the glifv3 first.

screenshot:
https://hsv.googleplex.com/5599282846498816
https://hsv.googleplex.com/5354235030929408

Bug: 233032365
Bug: 233036258

Change-Id: I7cdd25d34eee6dd8593c550ac803de5852ad870c