Calculate list type first in ManageApplications
Currently, list type is calculated in onCreate(), and will default to
LIST_TYPE_MAIN if the class name is not set to any special activity.

When adding new app list to onCreate(), but not changing
getSpaDestination() could led to issue when spa is on.

Calculate list type first in onAttach() as single source of truth could
solve this issue.

Bug: 235727273
Test: Manually with Settings with / without spa
Change-Id: If8a3d61cfd08d7d67e32253c3ac7266e5e9e7b4c