Add list preference for BT audio device type selection
Since BT devices do not populate the device type reliably we offer the
user the possibility to categorize the audio type of the selected
device. This is can be used by the AudioManager for enabling/disabling
the computation of sound dose.

Test: atest BluetoothDetailsAudioDeviceTypeControllerTest
Bug: 287011781

Merged-In: I797a92e1af4025596ef1c603ed4ab59813e3cbf0
Change-Id: I797a92e1af4025596ef1c603ed4ab59813e3cbf0