Fix pair other ear button not hidden correctly for dual device
If the device is a dual device which supports both ASHA + LE Audio, it'll pair the other ear with CSIP and group it as a member device. If LE Audio is not enabled, the current logic will fail to recognize the other ear is already paired. Update the logic to well handling this situation.

Bug: 296646732
Test: atest BluetoothDetailsPairOtherControllerTest
Change-Id: Iba9655488203a242ad0e87764d6bf4df595e977d