Disable the large screen optimization system property by default
When the property is true, Settings will invoke SplitController's APIs
to check if split support is available. However on a regular phone, it's
uncessary to take time to execute the APIs, which impacts the app
startup time.

Setting the property false by default can help skip the API invocation
when the property is not set. Developers can customize it in the build
file for large screen devices.

Test: build
Bug: 297794010
Change-Id: I91822d4b6afd49cd08fc546fbe5f231b2debc0f2