Don't add private profile accounts to Personal tab
If there are no tabs to show then the single Passwords & accounts page
shows all the profiles under the corresponding headings.
No tabs are shown only when the PS is locked and there's no work
profile. And if PS is locked then we don't want to show any PS setting.
On the other hand if PS is not locked then we'll have tabs.

Bug: 309423484
Test: manual
Change-Id: Iebc62fbc4caf21988af6aa1af4fbd570f4c80ada