Fix notification vibration demos in settings
While in zen mode, notification vibrations are ignored. These vibrations will also be ignored when attempting to demo the vibration intensity in settings.

Adding FLAG_BYPASS_INTERRUPTION_POLICY to the attributes of the vibration preview, to allow it to bypass the interruption policy and demo the vibration intensity as expected.

Fix: 301319836
Test: N/A
Change-Id: I240a896ff8e9516388a0238882db07f9680a184a