Add a preference in reset options to delete private space
This takses care to add an entry point in Settings->System->Reset
options->Delete private space when private space feature is supported on
the device. Entry point appears both when private space exists and also
when private is not created.
On selecting asks for user authentication and on successful
authentication using device lock shows a dialog asking for
confirmation and calls for deletion private space.

Based on this entrypoint it would not be possible to guess if private
space exists on the device.

Recording link : b/329041740#comment3

Bug: 329041740
Test: atest ResetOptionsDeletePrivateSpaceControllerTest and verified
this doesn't give away existence of private space

Change-Id: I9a5b908e7a3f9edaf609cf7e6a87f9842d689ce6