Register callback again after service is connected
The exception is thrown when trying to unregister a not-registered callback. This may happen when after rebooting the device, the proxy is not attached to the service while registerCallback() is called in onResume(). The callback will not be correctly registered and hence causing this exception when unregisterCallback() is called in onPause().

Add listener to listen the onServiceConnected() callback and re-register
the callback if the service is not connected when user enters to the
page.

Bug: 330116041
Test: manually test the scenario and confirm no exception thrown
Change-Id: I1f99d067e28e285bcbfff1f34bd322cdbac8063a