Try select preset independently after group operation failed
For some binaural models, when only one side is connected and call a group operation, the remote device will return ERROR_REMOTE_OPERATION_REJECTED. We can re-try to selectthe  preset for each device independently in this case. After the other device is reconnected, the preset should automatically sync in the remote side.

Bug: 300015207
Test: manual test with Gn Resound LE sample and verify it works
Test: atest BluetoothDetailsHearingAidsPresetsControllerTest
Change-Id: I3ea621577c248c974505b9ce65d8ef260f5b87e1