Clear settings task on deleting private space
When private space is deleted from private space settings launched from
the All apps settings cog the settings sub activity is not cleared and
visibile in recents.
With this change it is taken care to remove and clear the sub settings
task hosting the Private space settings page when private space does not
exist and to make sure the activity is removed if it is the last
activity in that stack.

Bug: 333538964
Test: Manual
Change-Id: I22ba8ca8d2947c6604280cf398cd47434038d542