Add flow for updating zen mode status on modes pages
The main responsibility for updating stored ZenMode data when updated is the Settings page, which reloads mode data whenever any changes are detected. This change adds a process for that Settings page to then update the ZenMode data associated with all controllers on that page.

Flag: android.app.modes_ui
Bug: 335259054
Test: manual with toy implementations
Change-Id: Ie1692da54d962c90378085d8bd8573e5b93f8eb6