Check config to Show/Hide Private space Hide Illustration
Enable/disable the search tile content and hide illustration
in Hide Private Space Settings page based on the config value.
By default the config value is set to true which will be used in Pixel
to show the illustration and search tile content.

OEM can set value of config_enableSearchTileHideIllustrationInPrivateSpace in the device
overlay file to false to disable it and with this change it is taken
care to hide them if the config value is false

Screenshot:
go/ss/56WxtgXd5NSqVFm.png

Bug: 346612477
Test: Manual
Flag: android.multiuser.private_space_search_illustration_config
Change-Id: Ia9fbdebd9a220f4d995369d662a75989f07a2916