Add interstitial layout for larger screen devices
For large enough screens, always use the portrait (all vertical in one column) layout for the intestitial, since the landscape split in two halves only makes sense for smaller screens.

Renames the layout to mode_interstitial_layout_base and uses an alias so that we can reuse the same layout but with additional padding.

Fixes: 361344824
Test: manual
Flag: android.app.modes_ui
Change-Id: I4e02e9a246ee799e3a79508f11290397a236574c