Right-align "end time" text on the schedule setter
When the "end time" text label is wider than the actual time, then aligning it with the start of the time causes it to wrap, even when there's plenty more space on the screen. This change right-aligns it in all cases (even when it's shorter), as it's hard to get all three of "right-aligned when too long", "left-aligned to the time display when short enough", and "wraps when it hits the center line instead of overflowing".

Fixes: 367054151
Test: manual, in English & Danish
Flag: android.app.modes_ui
Change-Id: I106864d6fb09aeac5fcc8d22e8d3d67b1b31ce4e