OnDeviceRecognition shows only profile group
Currently, if there are multiple users on the device, then the Settings
for OnDeviceRecognition show a chooser for all users on the device, with
only the users from the same profile group as the current user working.
It shouldn't even show the users from other profile groups; we fix that
here.

Bug: 361095700
Test: on a device with multiple users and profiles, made sure that only
users from the same profile group appeared
Flag: EXEMPT bugfix

Change-Id: I139e32a6a2cfede8d8a8abf573dcbaeede789f95