Show custom unlock title for private profile
With this change whenever private space is unlocked and has its own lock
different from device lock, it shows a custom title for PS in the unlock
screen.

Custom title for private profile will be  shown in the auth screen while
authentiation of private lock for the below usecase inside PS settings.
- Auth shown while Delete private space from PS settings
- Switching to device screen lock from a separate private space lock
- Changing the private space lock type(Ex: from pattern to PIN)
- Adding a biometric(face/fingerprint) for private profile

When auth screen with custom title is put to background and the screen
is in the recents list and if private space is locked then the auth
scren with the title is removed from recents list along with the other
settings task.

Bug: 323835257
Test: Manual
Change-Id: Ia0cc0b9bb54a429ebe027dc0937320e7fd23cb81
Flag: android.multiuser.show_custom_unlock_title_inside_private_profile