Implement the preference switch for minimalism on ls settings page
Implement the preference switch for notification minimalism on the
notifications on lock screen settings page. The illustrations and
summaries will also update when switching the preference. This
preference will hide when the main toggle is unset or when the lock
screen notification minimalism feature flag is disabled.

Bug: 367455695
Flag: com.android.server.notification.notification_lock_screen_settings
Test: adb shell settings get secure \
 lock_screen_notification_minimalism <1|0>
Change-Id: Ie6596e518bd1aa2e828587d7d174b57256aec45b